Split out subnetwork code
